Refactor app.rb into modules
